refactor: complete the parse side of the marshal seam
All 7 entities rendered through internal/marshal, but only 4 parsed
there. This moves the last three fs-resident parsers onto the seam, so
markdown -> entity now has one owner in both directions:

- marshal.MarkdownToStatusUpdate replaces parseUpdateContent (the hand
  line-scanner shared by project/initiative updates), rebuilt on
  marshal.Parse + ScalarToString. Contract preserved: default onTrack,
  case-insensitive normalization of the spelled variants, explicit
  unknown health -> FieldError, frontmatter-with-empty-body ->
  FieldError, no-frontmatter whitespace -> no-op for the callers.
- marshal.MarkdownToLabelUpdate / ParseNewLabel replace
  parseLabelMarkdown / parseNewLabelMarkdown / parseYAMLFrontmatter
  (the hand YAML scanner and its quote-strip hack are deleted).
- marshal.MarkdownToProjectEdit / MarkdownToInitiativeEdit are typed
  extraction structs (name via ScalarToString, lists via
  StringSliceFromYAML, labels as raw+present pair) — extraction only,
  no diffing: the diff keeps its owners (scalarEdit, labelsEdit,
  reconcileLinks). marshal.Document no longer appears in the
  project/initiative flush paths.

FieldError moves to marshal (fielderror.go) — the module that discovers
a bad field mints the error that names it. internal/fs keeps
`type FieldError = marshal.FieldError` (an alias), so every
construction site and the errors.As in classifyMutationErr match
unchanged.

scalarEdit's constructor now takes literal name/body strings; its
diff/trim/divergence logic is untouched and its tests got simpler.

Two recorded behavior changes, both loud-over-silent:
1. Unclosed frontmatter in a status update is now
   FieldError{frontmatter} -> EINVAL, instead of silently posting the
   raw bytes (delimiter and all) as the update body.
2. An unquoted hex color (`color: #FF0000`) parses in YAML as a
   comment (key present, nil value); both label parsers now reject it
   with quoting guidance instead of silently dropping the edit. The
   render side already single-quotes, so render -> parse stays a
   fixpoint.

Tests: parseUpdateContent table transferred to marshal wholesale plus
the two new behaviors; label parser tables transferred and extended
(nil-value vs key-absent, round-trip fixpoint asserts the quoted
render); render -> MarkdownTo*Edit identity pins for
project/initiative. CONTEXT.md gains the "Entity parse" sibling entry.
