+++ internal/fs/projects_test.go
@@ -1,7 +1,6 @@
 package fs
 
 import (
-	"errors"
 	"strings"
 	"testing"
 	"time"
@@ -79,150 +78,6 @@ func TestProjectDirName(t *testing.T) {
 			got := projectDirName(tt.project)
 			if got != tt.want {
 				t.Errorf("projectDirName() = %q, want %q", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestParseUpdateContent(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		name       string
-		content    string
-		wantBody   string
-		wantHealth string
-		wantField  string // non-empty: expect a *FieldError on this field
-	}{
-		{
-			name:       "plain text",
-			content:    "This is a status update.",
-			wantBody:   "This is a status update.",
-			wantHealth: "onTrack",
-		},
-		{
-			name: "with frontmatter onTrack",
-			content: `---
-health: onTrack
----
-All systems go!`,
-			wantBody:   "All systems go!",
-			wantHealth: "onTrack",
-		},
-		{
-			name: "with frontmatter atRisk",
-			content: `---
-health: atRisk
----
-Some delays expected.`,
-			wantBody:   "Some delays expected.",
-			wantHealth: "atRisk",
-		},
-		{
-			name: "with frontmatter offTrack",
-			content: `---
-health: offTrack
----
-Blocked by dependencies.`,
-			wantBody:   "Blocked by dependencies.",
-			wantHealth: "offTrack",
-		},
-		{
-			name: "health with spaces (on track)",
-			content: `---
-health: "on track"
----
-Update body`,
-			wantBody:   "Update body",
-			wantHealth: "onTrack",
-		},
-		{
-			name: "health with hyphens (at-risk)",
-			content: `---
-health: at-risk
----
-Body text`,
-			wantBody:   "Body text",
-			wantHealth: "atRisk",
-		},
-		{
-			name: "health with quotes",
-			content: `---
-health: 'off-track'
----
-Critical issues`,
-			wantBody:   "Critical issues",
-			wantHealth: "offTrack",
-		},
-		{
-			name:       "empty content",
-			content:    "",
-			wantBody:   "",
-			wantHealth: "onTrack",
-		},
-		{
-			name:       "whitespace only",
-			content:    "   \n\n   ",
-			wantBody:   "",
-			wantHealth: "onTrack",
-		},
-		{
-			name: "frontmatter without closing delimiter",
-			content: `---
-health: atRisk
-No closing delimiter`,
-			wantBody:   "---\nhealth: atRisk\nNo closing delimiter",
-			wantHealth: "onTrack",
-		},
-		{
-			name: "multiline body",
-			content: `---
-health: onTrack
----
-Line 1
-Line 2
-Line 3`,
-			wantBody:   "Line 1\nLine 2\nLine 3",
-			wantHealth: "onTrack",
-		},
-		{
-			name: "unknown health rejected, not coerced to onTrack",
-			content: `---
-health: critical
----
-Everything is on fire`,
-			wantField: "health",
-		},
-		{
-			name: "frontmatter with empty body rejected",
-			content: `---
-health: atRisk
----
-`,
-			wantField: "body",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			gotBody, gotHealth, err := parseUpdateContent([]byte(tt.content))
-			if tt.wantField != "" {
-				var ferr *FieldError
-				if !errors.As(err, &ferr) {
-					t.Fatalf("parseUpdateContent() err = %v, want *FieldError on %q", err, tt.wantField)
-				}
-				if ferr.Field != tt.wantField {
-					t.Errorf("parseUpdateContent() FieldError.Field = %q, want %q", ferr.Field, tt.wantField)
-				}
-				return
-			}
-			if err != nil {
-				t.Fatalf("parseUpdateContent() unexpected error: %v", err)
-			}
-			if gotBody != tt.wantBody {
-				t.Errorf("parseUpdateContent() body = %q, want %q", gotBody, tt.wantBody)
-			}
-			if gotHealth != tt.wantHealth {
-				t.Errorf("parseUpdateContent() health = %q, want %q", gotHealth, tt.wantHealth)
 			}
 		})
 	}
