+++ internal/fs/scalaredit.go
@@ -2,16 +2,14 @@ package fs
 
 import (
 	"strings"
-
-	"github.com/jra3/linear-fuse/internal/marshal"
 )
 
 // scalarEdit is the diff of the two scalar fields both project.md and
 // initiative.md expose for editing — a name (frontmatter) and a description
-// (body). It owns the change decision (what counts as "changed"), the name
-// coercion, and the read-your-writes divergence classification, so the two
-// handlers no longer each hand-roll a `fieldChanged` flag and a byte-identical
-// commitWriteBack compare closure. See CONTEXT.md "Scalar edit (scalarEdit)".
+// (body). It owns the change decision (what counts as "changed") and the
+// read-your-writes divergence classification, so the two handlers no longer
+// each hand-roll a `fieldChanged` flag and a byte-identical commitWriteBack
+// compare closure. See CONTEXT.md "Scalar edit (scalarEdit)".
 //
 // It stays neutral to the entity type: the caller maps name/desc onto its own
 // typed update input (api.ProjectUpdateInput / api.InitiativeUpdateInput) and
@@ -22,18 +20,18 @@ type scalarEdit struct {
 	origName, origDesc string  // pre-write values, for the divergence "original"
 }
 
-// newScalarEdit diffs a parsed document against the current name/description.
-// The body maps to the description; both sides are trimmed for the change test
-// so a render/parse trailing-newline delta doesn't read as an edit, and the
-// trimmed body is what we send. The frontmatter name is coerced the way the
-// issue handler coerces its scalars (a numeric or bare-scalar name updates
-// rather than being silently dropped); an empty or unchanged name is left alone.
-func newScalarEdit(doc *marshal.Document, curName, curDesc string) scalarEdit {
+// newScalarEdit diffs an already-parsed name/body against the current
+// name/description (marshal.MarkdownToProjectEdit/MarkdownToInitiativeEdit own
+// the extraction and the name coercion). The body maps to the description;
+// both sides are trimmed for the change test so a render/parse
+// trailing-newline delta doesn't read as an edit, and the trimmed body is what
+// we send. An empty or unchanged name is left alone.
+func newScalarEdit(name, body string, curName, curDesc string) scalarEdit {
 	e := scalarEdit{origName: curName, origDesc: curDesc}
-	if newDesc := strings.TrimSpace(doc.Body); newDesc != strings.TrimSpace(curDesc) {
+	if newDesc := strings.TrimSpace(body); newDesc != strings.TrimSpace(curDesc) {
 		e.desc = &newDesc
 	}
-	if name := marshal.ScalarToString(doc.Frontmatter["name"]); name != "" && name != curName {
+	if name != "" && name != curName {
 		e.name = &name
 	}
 	return e
