+++ internal/marshal/project.go
@@ -9,11 +9,12 @@ import (
 // ProjectToMarkdown renders the editable-only project.md: name, initiatives,
 // labels, and the description body. Server-managed fields live in project.meta
 // (see ProjectMetaToMarkdown), so a successful write never rewrites the bytes
-// the writer wrote. The parse side is scalarEdit (name/description) plus
-// reconcileLinks (the initiatives list) plus resolveProjectLabels (the labels
-// list) in internal/fs. labelNames is the project's labelIds mapped to catalog
-// names by the caller — an unknown ID arrives verbatim (round-trip invariant);
-// the key is omitted when empty (delete-the-line clears).
+// the writer wrote. The parse side is MarkdownToProjectEdit below; the diffs
+// stay with internal/fs's scalarEdit (name/description), reconcileLinks (the
+// initiatives list), and labelsEdit (the labels list). labelNames is the
+// project's labelIds mapped to catalog names by the caller — an unknown ID
+// arrives verbatim (round-trip invariant); the key is omitted when empty
+// (delete-the-line clears).
 func ProjectToMarkdown(project *api.Project, labelNames []string) ([]byte, error) {
 	fm := map[string]any{"name": project.Name}
 
@@ -60,4 +61,37 @@ func ProjectMetaToMarkdown(project *api.Project) ([]byte, error) {
 		fm["targetDate"] = *project.TargetDate
 	}
 	return Render(&Document{Frontmatter: fm})
+}
+
+// ProjectEdit is what an edited project.md says — extraction and coercion
+// only, no diffing (the diff has owners: scalarEdit for name/body, labelsEdit
+// for labels, reconcileLinks for initiatives). Labels keep their raw
+// value + presence pair because labelsEdit downstream owns the label coercion
+// (ID passthrough, ambiguity); initiatives collapse to a plain slice where
+// absent ⇒ empty, today's unlink-all semantics.
+type ProjectEdit struct {
+	Name          string
+	Body          string
+	LabelsRaw     any
+	LabelsPresent bool
+	Initiatives   []string
+}
+
+// MarkdownToProjectEdit parses an edited project.md into its editable field
+// set. The name is coerced via ScalarToString (a numeric/bare-scalar name
+// arrives as its string form, not a silent drop); the body passes through
+// verbatim for scalarEdit's trim-aware diff.
+func MarkdownToProjectEdit(content []byte) (*ProjectEdit, error) {
+	doc, err := Parse(content)
+	if err != nil {
+		return nil, err
+	}
+	rawLabels, labelsPresent := doc.Frontmatter["labels"]
+	return &ProjectEdit{
+		Name:          ScalarToString(doc.Frontmatter["name"]),
+		Body:          doc.Body,
+		LabelsRaw:     rawLabels,
+		LabelsPresent: labelsPresent,
+		Initiatives:   StringSliceFromYAML(doc.Frontmatter["initiatives"]),
+	}, nil
 }
